+++ package.json
@@ -4,7 +4,8 @@
   "private": true,
   "homepage": "https://studentiunimi.it/",
   "dependencies": {
-    "@fluentui/react": "8.30.1",
+    "@fluentui/merge-styles": "^8.5.3",
+    "@fluentui/react": "^8.90.0",
     "@fluentui/react-hooks": "^8.2.4",
     "@fluentui/react-theme-provider": "^0.18.4",
     "@fortawesome/fontawesome-svg-core": "^1.2.36",
