+++ pages/organization.tsx
@@ -3,14 +3,19 @@ import { Text, Link } from 'office-ui-fabric-react';
 import { Container } from 'react-bootstrap';
 import { Persona, PersonaSize } from 'office-ui-fabric-react/lib-commonjs/Persona';
 import { useTheme } from '@fluentui/react-theme-provider';
-import { getNetworkMembers } from '../services/Requests';
-import { CompoundButton } from '@fluentui/react/lib-commonjs/Button';
+import { getNetworkMembers } from '../src/services/Requests';
+//import { CompoundButton } from '@fluentui/react/lib-commonjs/Button';
 import { IIconProps } from '@fluentui/react';
 import { Image } from 'office-ui-fabric-react/lib-commonjs/Image';
-import { semibold } from '../services/Fonts';
+import { semibold } from '../src/services/Fonts';
 import Col from 'react-bootstrap/Col';
 import Row from 'react-bootstrap/Row';
-import LocalizationService from "../services/LocalizationService";
+import LocalizationService from "../src/services/LocalizationService";
+import Document, { Head, Html, Main, NextScript } from 'next/document';
+import { Stylesheet } from '@fluentui/merge-styles';
+import { resetIds } from '@fluentui/react';
+
+const stylesheet = Stylesheet.getInstance();
 
 const developers: any = [
     { name: "Giuseppe Del Campo", description: { it: "Sviluppatore del sito web", en: 'Website Developer' }, username: 'Giuseppetm', user_id: 597678134, github: "https://github.com/Giuseppetm", website: "https://giuseppetm.netlify.app/" },
@@ -25,6 +30,8 @@ const OrganizationView = () => {
     const networkMembers = getNetworkMembers();
     const icon: IIconProps = { iconName: 'DocumentSearch' };
     const imageProperties = { display: 'inline-block', width: '100%' };
+    
+    resetIds();
 
     return (
         <div className="organization pb-3">
@@ -53,9 +60,11 @@ const OrganizationView = () => {
                             </div>
 
                             <div className="mb-2">
+                                {/*
                                 <CompoundButton primary theme={theme} secondaryText={locale?.aboutUs.button.text2} href="https://github.com/StudentiUniMi/docs/raw/main/statuto.pdf" style={{ textDecoration: 'none', boxShadow: theme.effects.elevation8 }} iconProps={icon}>
                                     {locale?.aboutUs.button.text1}
                                 </CompoundButton>
+                                */}
                             </div>
                         </Col>
                     </Row>
