+++ pages/services.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Col from 'react-bootstrap/Col';
 import Row from 'react-bootstrap/Row';
-import LocalizationService from "../services/LocalizationService";
+import LocalizationService from "../src/services/LocalizationService";
 import JsxParser from 'react-jsx-parser';
 import { useHistory } from 'react-router-dom';
 import { FontSizes } from '@fluentui/theme';
@@ -10,8 +10,8 @@ import { Container } from 'react-bootstrap';
 import { useTheme } from '@fluentui/react-theme-provider';
 import { DocumentCardPreview, IDocumentCardPreviewProps } from 'office-ui-fabric-react/lib-commonjs/DocumentCard';
 import { Card, ICardTokens } from '@uifabric/react-cards';
-import { semibold } from '../services/Fonts';
-import { getRedirects, getGuides, getTools } from '../services/Requests';
+import { semibold } from '../src/services/Fonts';
+import { getRedirects, getGuides, getTools } from '../src/services/Requests';
 import { Pivot, PivotItem } from '@fluentui/react';
 import { Image } from 'office-ui-fabric-react/lib-commonjs/Image';
 import { Link } from '@fluentui/react';
