+++ src/index.tsx
@@ -5,7 +5,7 @@ import "swiper/components/pagination/pagination.min.css";
 import "swiper/components/navigation/navigation.min.css";
 import 'bootstrap/dist/css/bootstrap.min.css';
 import './index.scss';
-import MainView from '../pages';
+import MainView from '../pages/mainview';
 
 ReactDOM.render(
   <React.StrictMode>
